Move bench to separate folder
