+++ src/daft-minhash/src/minhash.rs
@@ -6,8 +6,6 @@ use std::{
 use common_error::DaftResult;
 use mur3::murmurhash3_x86_32;
 
-extern crate test;
-
 const MERSENNE_PRIME: u64 = (1 << 61) - 1;
 const MAX_HASH: u32 = 0xffffffff;
 
@@ -126,14 +124,6 @@ pub fn minhash(s: &str, num_hashes: usize, ngram_size: usize, seed: u32) -> Daft
 #[cfg(test)]
 mod tests {
     use super::*;
-    use std::ops::Range;
-    use test::Bencher;
-
-    const N_TOKENS: usize = 10000;
-    const N_CHARS: Range<usize> = 1..20;
-
-    const NUM_HASHES: usize = 128;
-    const NGRAM_SIZE: usize = 13;
 
     #[test]
     fn test_rem_min() {
@@ -176,21 +166,5 @@ mod tests {
         for i in 0..16 {
             assert!(res2[i] == res3[i]);
         }
-    }
-
-    #[bench]
-    fn bench_minhash(b: &mut Bencher) {
-        let mut rng = fastrand::Rng::with_seed(42);
-        let mut s: String = String::new();
-        for i in 0..N_TOKENS {
-            if i > 0 {
-                s.push(' ');
-            }
-            let s_chars = rng.usize(N_CHARS);
-            for _ in 0..s_chars {
-                s.push(rng.alphanumeric());
-            }
-        }
-        b.iter(|| minhash(&s, NUM_HASHES, NGRAM_SIZE, 1));
     }
 }
